refactored to add dialog via dialog.js rather than criminal.js to contain the dialog element and event listeners together for readability.
