+++ scripts/main.js
@@ -1,8 +1,9 @@
 import { getCriminals } from "./criminals/CriminalProvider.js";
-import { criminalList, initializeCloseButtonEvents } from "./criminals/CriminalList.js";
+import { criminalList } from "./criminals/CriminalList.js";
 import ConvictionSelect from "./convictions/ConvictionsSelect.js";
 import { getConvictions } from "./convictions/ConvictionsProvider.js";
 import { NoteForm } from "./notes/NoteForm.js";
+import { initializeCloseButtonEvents } from "./criminals/dialog.js";
 
 NoteForm();
 getCriminals().then(criminalList).then(initializeCloseButtonEvents);
